refactor: reorganize processors
