+++ metrics/csv_to_aligned_txt.py
